+++ detectors/node/opentelemetry-resource-detector-aws/src/detectors/AwsEc2DetectorSync.ts
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+import { context } from '@opentelemetry/api';
+import { suppressTracing } from '@opentelemetry/core';
 import {
   DetectorSync,
   IResource,
@@ -56,7 +58,10 @@ class AwsEc2DetectorSync implements DetectorSync {
   readonly MILLISECOND_TIME_OUT = 5000;
 
   detect(_config?: ResourceDetectionConfig): IResource {
-    return new Resource({}, this._getAttributes());
+    const attributes = context.with(suppressTracing(context.active()), () =>
+      this._getAttributes()
+    );
+    return new Resource({}, attributes);
   }
 
   /**