+++ plugins/node/instrumentation-tedious/test/api.ts
@@ -16,14 +16,14 @@
 
 import * as assert from 'assert';
 import { promisify } from 'util';
-import type { Connection, Request, TYPES, ConnectionConfig } from 'tedious';
+import type { Connection, Request, TYPES } from 'tedious';
 
 type Method = keyof Connection & ('execSql' | 'execSqlBatch' | 'prepare');
 export type tedious = {
   Connection: typeof Connection;
   Request: typeof Request;
   TYPES: typeof TYPES;
-  ConnectionConfig: ConnectionConfig;
+  ConnectionConfig: any;
 };
 
 export const makeApi = (tedious: tedious) => {
@@ -32,7 +32,7 @@ export const makeApi = (tedious: tedious) => {
     return `[dbo].[${resource}]`;
   };
 
-  const createConnection = (config: ConnectionConfig): Promise<Connection> => {
+  const createConnection = (config: any): Promise<Connection> => {
     return new Promise((resolve, reject) => {
       const connection = new tedious.Connection(config);
 
@@ -268,11 +268,11 @@ export const makeApi = (tedious: tedious) => {
     },
     execute: (connection: Connection): Promise<number> => {
       return new Promise((resolve, reject) => {
-        const requestDoneCb = (err: any, rowCount: number) => {
+        const requestDoneCb = (err: any, rowCount?: number) => {
           if (err) {
             return reject(err);
           }
-          resolve(rowCount);
+          resolve(rowCount!);
         };
         // <2.2.0 didn't take bulkOptions
         const request =
@@ -294,7 +294,7 @@ export const makeApi = (tedious: tedious) => {
           // required in <=11.5. not supported in 14
           request.addRow({ c1: 1 });
           request.addRow({ c1: 2, c2: 'hello' });
-          return connection.execBulkLoad(request);
+          return (connection.execBulkLoad as any)(request);
         }
 
         (connection.execBulkLoad as any)(request, [