+++ plugins/node/instrumentation-tedious/test/instrumentation.test.ts
@@ -37,7 +37,7 @@ import {
 import * as assert from 'assert';
 import { TediousInstrumentation } from '../src';
 import makeApi from './api';
-import type { Connection, ConnectionConfig } from 'tedious';
+import type { Connection } from 'tedious';
 import * as semver from 'semver';
 
 const port = Number(process.env.MSSQL_PORT) || 1433;
@@ -50,7 +50,7 @@ const instrumentation = new TediousInstrumentation();
 instrumentation.enable();
 instrumentation.disable();
 
-const config: ConnectionConfig & { userName: string; password: string } = {
+const config: any = {
   userName: user,
   password,
   server: host,
@@ -79,7 +79,11 @@ const incompatVersions =
   (semver.lt(processVersion, '15.0.0') &&
     semver.gte(tediousVersion, '16.0.0')) ||
   // tedious@17 removed support for node v16 and v19 https://github.com/tediousjs/tedious/releases/tag/v17.0.0
-  (semver.lt(processVersion, '17.0.0') && semver.gte(tediousVersion, '17.0.0'));
+  (semver.lt(processVersion, '17.0.0') &&
+    semver.gte(tediousVersion, '17.0.0')) ||
+  // tedious@19 removed support for node <18.17.0 https://github.com/tediousjs/tedious/releases/tag/v19.0.0
+  (semver.lt(processVersion, '18.17.0') &&
+    semver.gte(tediousVersion, '19.0.0'));
 
 describe('tedious', () => {
   let tedious: any;