+++ plugins/node/opentelemetry-instrumentation-express/src/utils.ts
@@ -45,11 +45,33 @@ export const storeLayerPath = (request: PatchedRequest, value?: string) => {
 };
 
 /**
+ * Recursively search the router path from layer stack
+ * @param path The path to reconstruct
+ * @param layer The layer to reconstruct from
+ * @returns The reconstructed path
+ */
+export const getRouterPath = (path: string, layer: ExpressLayer): string => {
+  const stackLayer = layer.handle?.stack?.[0];
+
+  if (stackLayer?.route?.path) {
+    return `${path}${stackLayer.route.path}`;
+  }
+
+  if (stackLayer?.handle?.stack) {
+    return getRouterPath(path, stackLayer);
+  }
+
+  return path;
+};
+
+/**
  * Parse express layer context to retrieve a name and attributes.
+ * @param route The route of the layer
  * @param layer Express layer
  * @param [layerPath] if present, the path on which the layer has been mounted
  */
 export const getLayerMetadata = (
+  route: string,
   layer: ExpressLayer,
   layerPath?: string
 ): {
@@ -57,20 +79,26 @@ export const getLayerMetadata = (
   name: string;
 } => {
   if (layer.name === 'router') {
+    const maybeRouterPath = getRouterPath('', layer);
+    const extractedRouterPath = maybeRouterPath
+      ? maybeRouterPath
+      : layerPath || route || '/';
+
     return {
       attributes: {
-        [AttributeNames.EXPRESS_NAME]: layerPath,
+        [AttributeNames.EXPRESS_NAME]: extractedRouterPath,
         [AttributeNames.EXPRESS_TYPE]: ExpressLayerType.ROUTER,
       },
-      name: `router - ${layerPath}`,
+      name: `router - ${extractedRouterPath}`,
     };
   } else if (layer.name === 'bound dispatch') {
     return {
       attributes: {
-        [AttributeNames.EXPRESS_NAME]: layerPath ?? 'request handler',
+        [AttributeNames.EXPRESS_NAME]:
+          (route || layerPath) ?? 'request handler',
         [AttributeNames.EXPRESS_TYPE]: ExpressLayerType.REQUEST_HANDLER,
       },
-      name: `request handler${layer.path ? ` - ${layerPath}` : ''}`,
+      name: `request handler${layer.path ? ` - ${route || layerPath}` : ''}`,
     };
   } else {
     return {
@@ -159,11 +187,13 @@ export const asErrorAndMessage = (
 export const getLayerPath = (
   args: [LayerPathSegment | LayerPathSegment[], ...unknown[]]
 ): string | undefined => {
-  if (Array.isArray(args[0])) {
-    return args[0].map(arg => extractLayerPathSegment(arg) || '').join(',');
+  const firstArg = args[0];
+
+  if (Array.isArray(firstArg)) {
+    return firstArg.map(arg => extractLayerPathSegment(arg) || '').join(',');
   }
 
-  return extractLayerPathSegment(args[0]);
+  return extractLayerPathSegment(firstArg);
 };
 
 const extractLayerPathSegment = (arg: LayerPathSegment) => {