+++ plugins/node/instrumentation-fs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.15.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-fs-v0.14.0...instrumentation-fs-v0.15.0) (2024-09-02)
+
+
+### Features
+
+* update deps matching "@opentelemetry/" ([9fa058e](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/9fa058ebb919de4e2a4e1af95b3c792c6ea962ac))
+
 ## [0.14.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-fs-v0.13.0...instrumentation-fs-v0.14.0) (2024-07-03)